working on some rendering wierdness with socket
