+++ src/views/CombatMap/index.tsx
@@ -8,22 +8,14 @@ import Map from "./components/Map";
 import { Typography } from "@components/Typography/Typography";
 import { useDeepCompareEffectNoCheck } from "use-deep-compare-effect";
 import { Spacer } from "@components/Spacer/Spacer";
-import { Button } from "@components/Button/Button";
 import ExtraTokenContent from "./components/Token/ExtraTokenContent";
-import SettingsDrawer from "./components/SettingsDrawer";
 import CharacterTokenContent from "./components/Token/CharacterTokenContent";
 import useCombatMapStore from "./CombatMapStore";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faEraser, faMinus, faPlus } from "@fortawesome/free-solid-svg-icons";
 import { useCombatMap, useCombatMapSocketService, useCombatMapTokens, useMutateCombatToken, useUpdateCombatMap } from "@services/CombatMapService";
 import { FullScreen, useFullScreenHandle } from "react-full-screen";
-import FloatingButtonContainer from "@components/FloatingButtonContainer";
 import Spinner from "@components/Spinner";
-import TokensDrawer from "./components/TokensDrawer";
-import { useSpring, animated } from '@react-spring/web';
-import ColorPicker from "@components/ColorPicker/ColorPicker";
 import { useCampaignCharacters } from "@services/CharacterService";
-import { TextInput } from "@components/TextInput/TextInput";
+import UtilButtons from "./components/UtilButtons";
 type DroppableToken = {
   id: string,
   data: any,
@@ -36,12 +28,7 @@ const CombatMap = ({combatIdOverride = "", isPlayer = false}) => {
     const [searchParams, setSearchParams] = useSearchParams();
     const scale = Number(searchParams.get("scale")) || 1;
     const drawing = String(searchParams.get("drawing")) || "";
-    const color = String(searchParams.get("color")) || "black";
-    const drawSize = Number(searchParams.get("drawSize")) || 0;
     const eraserOn = searchParams.get("eraserOn") === "on" || false;
-
-    const [isSettingsDrawerOpen, setIsSettingsDrawerOpen] = useState(false);
-    const [isTokenDrawerOpen, setIsTokenDrawerOpen] = useState(false);
 
     const [tokens, setTokens] = useState<DroppableToken[]>([]);
     const [extraTokens, setExtraTokens] = useState<DroppableToken[]>([]);
@@ -50,18 +37,13 @@ const CombatMap = ({combatIdOverride = "", isPlayer = false}) => {
     const [altHeld, setAltHeld] = useState(false);
 
     const {mutate: mutateCombatToken, isLoading: isMutating} = useMutateCombatToken()
-    const { combat, isLoading, isRefetching } = useCombat(combatId);
-    const { combatMap, isLoading: isMapLoading, isRefetching: isMapRefetching } = useCombatMap(combatId || combatIdOverride);
+    const { combat, isLoading } = useCombat(combatId);
+    const { combatMap, isLoading: isMapLoading } = useCombatMap(combatId || combatIdOverride);
     const {mutate: editMap} = useUpdateCombatMap(combatMap?.docId || "");
     const {combatCharacters = [], isLoading: isCharactersLoading} = useCombatCharacters(combatId);
     const {tokens: combatTokens, isLoading: isTokensLoading} = useCombatMapTokens(combatMap?.docId || "", isMutating);
     const { characters: campaignCharacters = [] } = useCampaignCharacters(combat?.campaignDocId || "");
-    const [isHovered, setHovered] = useState(false);
 
-    const springs = useSpring({
-        translateY: isHovered ? -40 : 0,
-        opacity: isHovered ? 1 : 0
-    })
     const { currentTurnIndex } = combat;
 
     const mapRef = useRef<HTMLDivElement>(null); 
@@ -191,8 +173,7 @@ const CombatMap = ({combatIdOverride = "", isPlayer = false}) => {
 
     // set local data from remote
     useDeepCompareEffectNoCheck(() => {
-      if (!isMapLoading && !isMapRefetching && !isLoading && !isRefetching && !isMutating) {
-        console.log("Update from DB")
+      if (!isMapLoading && !isLoading && !isMutating) {
         const tokensFromCombat = combatCharacters?.map((character, index) => {
           const isPC = character?.playerDocId;
           //clean up types here
@@ -300,8 +281,6 @@ const CombatMap = ({combatIdOverride = "", isPlayer = false}) => {
       }))}
     );
 
-    
-
     const handleDragMove = (event: any) => {      
       const x = (selectedToken?.data?.position.x + event?.delta?.x/scale);
       const y = (selectedToken?.data?.position.y + event?.delta?.y/scale);
@@ -309,22 +288,8 @@ const CombatMap = ({combatIdOverride = "", isPlayer = false}) => {
       updateTokenPosition(event.active.id, x, y)
     }
 
-
-
     if (isLoading || isMapLoading || isCharactersLoading || isTokensLoading) {
       return <Spinner />
-    }
-
-    const getDrawingTitle = () => {
-      let title = "Drawing Off";
-      if (drawing === "drawing") {
-        title = "Drawing";
-      }
-      if (drawing === "fogOfWar") {
-        title = "Fog Of War"
-      }
-
-      return title;
     }
 
     return (
@@ -378,83 +343,12 @@ const CombatMap = ({combatIdOverride = "", isPlayer = false}) => {
               </Map>
             </DndContext>
           </div>
-          <FloatingButtonContainer>
-            <div className={css.buttonsContainer}>
-              {!mapContainer.active && !isPlayer && <Button onClick={() => setIsSettingsDrawerOpen(true)} color="secondary" animatedHover><Typography color="light">Settings</Typography></Button>}
-              {!mapContainer.active && <Button onClick={() => setIsTokenDrawerOpen(true)} color="secondary" animatedHover><Typography color="light">Tokens</Typography></Button>}
-              { !isPlayer && 
-                <div onMouseEnter={() => setHovered(true)} onMouseLeave={() => setHovered(false)}>
-                    <div style={{position: "relative", marginBottom: -50, zIndex: 2, display: "flex"}}>
-                      <Button onClick={() => setSearchParams(searchParams => {
-                          const nextState = drawing === "drawing" ? "fogOfWar" : drawing === "fogOfWar" ? "" : "drawing";
-
-                          searchParams.set("drawing", nextState);
-                          return searchParams;
-                        })
-                      } color={drawing ? "tertiary" : "secondary"} animatedHover><Typography color="light">{getDrawingTitle()}</Typography></Button>
-                      <Button onClick={() => setSearchParams(searchParams => {
-
-                          searchParams.set("eraserOn", eraserOn ? "off" : "on");
-                          return searchParams;
-                        })
-                      } color={eraserOn ? "success" : "error"} animatedHover><Typography color="light">{<FontAwesomeIcon icon={faEraser} />}</Typography></Button>
-                    </div>
-                    <animated.div style={springs}>
-                      <div className={css.hoveredDrawButtons}>
-                        <ColorPicker width={50} value={color} onChange={(value) => {
-                          setSearchParams(searchParams => {
-                            searchParams.set("color", String(value));
-                            return searchParams;
-                          })
-                        }}/>
-                        <TextInput className={css.drawSizeInput} number placeholder="Size" value={drawSize} onChange={(value) => setSearchParams(searchParams => {
-                            searchParams.set("drawSize", String(value));
-                            return searchParams;
-                          })
-                        } />
-                      </div>
-                    </animated.div>
-                </div>
-              }
-
-
-              <Button animatedHover onClick={() => mapContainer.active ? mapContainer.exit() : mapContainer.enter()}><Typography color="light">FullScreen</Typography></Button>
-              <div style={{display: "flex", justifyContent: "center", alignItems: "center", columnGap: 4}}>
-                <Button onClick={
-                    () => setSearchParams(searchParams => {
-                      searchParams.set("scale", String((scale - 0.1).toFixed(2)));
-                      return searchParams;
-                    })
-                  }><FontAwesomeIcon icon={faMinus} /></Button>
-                  <Typography weight="bold" color="light">Zoom</Typography>
-                <Button onClick={
-                    () => setSearchParams(searchParams => {
-                      searchParams.set("scale", String((scale + 0.1).toFixed(2)));
-                      return searchParams;
-                    })
-                  }><FontAwesomeIcon icon={faPlus} /></Button>
-              </div>
-              { isPlayer && 
-                <div>
-                  <Button animatedHover onClick={() => setSearchParams(searchParams => {
-                    searchParams.set("tab", "initiative");
-                    return searchParams
-                  })}><Typography>Health</Typography></Button>
-                </div>
-              }
-            </div>
-          </FloatingButtonContainer>
-            <SettingsDrawer
-              isOpen={isSettingsDrawerOpen}
-              onClose={() => setIsSettingsDrawerOpen(false)}
-              combatId={combatId}
-            />
-            <TokensDrawer
-              isOpen={isTokenDrawerOpen}
-              onClose={() => setIsTokenDrawerOpen(false)}
-              combatId={combatId}
-              isPlayer={isPlayer}
-            />
+          
+          <UtilButtons
+            isPlayer={isPlayer}
+            combatId={combatId}
+            mapContainer={mapContainer}
+          />
         </FullScreen>
       </div>
     );
